restructure utils, moved addresses to constants, seperated dev dependencies, cleanup truffle config, added env vars
