+++ .env.public
@@ -1,0 +1,2 @@
+INFURA_KEY=
+MNEMONIC=
