+++ package.json
@@ -9,26 +9,23 @@
   "dependencies": {
     "@openzeppelin/contracts": "^3.4.2-solc-0.7",
     "@openzeppelin/contracts-upgradeable": "^3.4.2-solc-0.7",
-    "@truffle/hdwallet-provider": "^1.2.1",
-    "@umaprotocol/truffle-ledger-provider": "^1.0.3",
-    "@uniswap/sdk": "^3.0.3",
-    "@uniswap/sdk-core": "^3.0.1",
     "@uniswap/v2-core": "^1.0.1",
     "@uniswap/v2-periphery": "^1.1.0-beta.0",
-    "@uniswap/v3-periphery": "^1.4.1",
-    "@uniswap/v3-sdk": "^3.8.2",
-    "bignumber.js": "^9.0.2",
-    "ganache-cli": "^6.12.1",
-    "ganache-core": "^2.13.1",
-    "truffle-hdwallet-provider": "^1.0.17",
-    "truffle-ledger-provider": "^1.0.3",
-    "web3": "^1.3.0"
+    "@uniswap/v3-periphery": "^1.4.1"
   },
   "devDependencies": {
     "@openzeppelin/test-helpers": "^0.5.10",
-    "chai-bn": "^0.2.1",
+    "@truffle/hdwallet-provider": "^2.0.8",
+    "@uniswap/sdk": "^3.0.3",
+    "@uniswap/sdk-core": "^3.0.1",
+    "@uniswap/v3-sdk": "^3.8.2",
+    "bignumber.js": "^9.0.2",
+    "dotenv": "^16.0.1",
+    "ganache-cli": "^6.12.1",
+    "ganache-core": "^2.13.1",
     "solidity-coverage": "^0.7.13",
-    "truffle-flattener": "^1.5.0"
+    "truffle-flattener": "^1.5.0",
+    "web3": "^1.7.3"
   },
   "scripts": {
     "start": "ganache-cli --fork https://eth-mainnet.alchemyapi.io/v2/oPuANrR8LdckQq_jyNsUxtlV76MHpMkm  --unlock 0x3675D2A334f17bCD4689533b7Af263D48D96eC72 --unlock 0xe8eA8bAE250028a8709A3841E0Ae1a44820d677b --unlock 0xfbb1b73c4f0bda4f67dca266ce6ef42f520fbb98 --unlock 0xfD88D7E737a06Aa9c62B950C1cB5eE63DA379AFd --unlock 0xb0aA1f98523Ec15932dd5fAAC5d86e57115571C7 --unlock 0xD6dABBc2b275114a2366555d6C481EF08FDC2556 --networkId 4447",
