+++ utilities/exchanges/uniswapV3/swap.js
@@ -3,7 +3,7 @@ const {
   abi: ISwapRouter,
 } = require("@uniswap/v3-periphery/artifacts/contracts/interfaces/ISwapRouter.sol/ISwapRouter.json");
 const ERC20  = require('@openzeppelin/contracts/build/contracts/ERC20.json')
-const addresses = require('../../migrations/addresses').development
+const addresses = require('../../../constants/addresses').development
 
 const swap = async (tokenIn, tokenOut, fee, toAddress, amountIn) => {
